[MIG] base_import_async: replace connector by queue_job
