+++ base_import_async/static/src/xml/import.xml
@@ -4,9 +4,9 @@
             <div title="When checked, the import will be executed as a background job, 
 after splitting your file in small chunks that will be processed independently. 
 Use this to import very large files.">
-                <input type="checkbox" class="oe_import_connector"
-                       id="oe_import_connector"/>
-                <label for="oe_import_connector">Import in the background</label>
+                <input type="checkbox" class="oe_import_queue"
+                       id="oe_import_queue"/>
+                <label for="oe_import_queue">Import in the background</label>
             </div>
         </t>
     </t>
