+++ oca_dependencies.txt
@@ -1,1 +1,2 @@
 connector
+queue
